Crypto farming implemented

+++ BattleCity/src/battlecity/game/Item.java
@@ -16,8 +16,8 @@ public abstract class Item implements Runnable, ItemInterface {
     private int newX;
     private int newY;
     private int life;
-    private int speedX;
-    private int speedY;
+    private float speedX;
+    private float speedY;
     private volatile long lastUpdateTime;
 
     private Orientation orientation;
@@ -98,11 +98,11 @@ public abstract class Item implements Runnable, ItemInterface {
         return life;
     }
 
-    public synchronized int getSpeedX() {
+    public synchronized float getSpeedX() {
         return speedX;
     }
 
-    public synchronized int getSpeedY() {
+    public synchronized float getSpeedY() {
         return speedY;
     }
 
@@ -133,11 +133,11 @@ public abstract class Item implements Runnable, ItemInterface {
         this.life = life;
     }
 
-    public synchronized void setSpeedX(int speed) {
+    public synchronized void setSpeedX(float speed) {
         this.speedX = speed;
     }
 
-    public synchronized void setSpeedY(int speed) {
+    public synchronized void setSpeedY(float speed) {
         this.speedY = speed;
     }
